requirements.txt
# ColPali + RAG dependencies
byaldi>=0.0.6
colpali-engine>=0.3.1
torch>=2.1.0
torchvision>=0.16.0
# before
torch>=2.1.0
transformers==4.44.2

# PDF processing
pdf2image>=1.16.3
Pillow>=10.0.0
pypdf>=3.17.0

# Vector store
faiss-cpu>=1.7.4
# Use faiss-gpu if CUDA available:
# faiss-gpu>=1.7.4

# VLM generation
groq>=0.9.0

# App & utilities
streamlit>=1.32.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation
scikit-learn>=1.3.0